requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1

# Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
requests==2.31.0
lxml==4.9.3

# AI and NLP (FREE - Google Gemini)
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.26.2

# Database
sqlalchemy==2.0.23

# Scheduling
apscheduler==3.10.4

# Email
secure-smtplib==0.1.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1

# Logging
colorlog==6.8.0

# Data Processing
pandas==2.1.4

# Optional: For better text processing
nltk==3.8.1
spacy==3.7.2

# Optional: For resume parsing
pdfplumber==0.10.3
python-docx==1.1.0
